Fix a misnested `.find(...)` expression for CG lookup
The misnesting caused the case with a trailing slash to not work, so
update the test such that it would have failed without this fix.

Also slightly reorganize lib/validator.js by initializing this closer
to where they are used. This makes it clear that the clobbering
comments were wrong, `groups` is actually just written in both the
if-branch and the else-branch.