Use @octokit/core + @octokit/plugin-throttling for GraphQL
Automatic throttling and retries should make this more reliable, and
failing queries are no longer caught and logged and instead allowed to
be fatal.

The use of GraphQL variables for cursor makes some of the pagination
code a bit more straightforward.

All queries are serialized, but by requesting 100 labels the overall
time is about the same as this avoid a lot of extra label queries.

Also normalize the indentation of the GraphQL query blocks.